+++ lib/body/body.dart
@@ -4,9 +4,7 @@ library;
 import 'package:flutter/material.dart';
 import 'package:flutter_muscle_anatomy/core/core.dart';
 import 'dart:math' as math;
-import 'package:path_drawing/path_drawing.dart';
 import 'package:collection/collection.dart';
-import 'package:easy_localization/easy_localization.dart';
 
 part 'body_mixins.dart' ;