+++ lib/core/core.dart
@@ -5,7 +5,6 @@ import 'package:flutter_muscle_anatomy/core/assets/svg_assets.dart';
 import 'package:flutter_muscle_anatomy/core/utils.dart';
 import 'package:path_drawing/path_drawing.dart';
 import 'package:xml/xml.dart';
-import 'package:easy_localization/easy_localization.dart';
 
 part 'localization.dart';