+++ lib/flutter_muscle_anatomy.dart
@@ -1,4 +1,13 @@
 library;
 
 export 'body/body.dart' show Male, Female, MuscleAnatomy, Anatomy;
-export 'core/core.dart' show Muscle, MusclePosition, BodyView;
+export 'core/core.dart'
+    show
+        Muscle,
+        MusclePosition,
+        BodyView,
+        MuscleLocalization,
+        BodyViewLocalization,
+        MusclePositionLocalization,
+        GenderLocalization,
+        MuscleAnatomyLocalization;