+++ test/flutter_muscle_anatomy_test.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_muscle_anatomy/core/core.dart';
 import 'package:flutter_muscle_anatomy/flutter_muscle_anatomy.dart';
 import 'package:flutter_test/flutter_test.dart';